feat(refactor): scuffle project refactor
This commit refactors and cleans up the project getting ready for our
launch in 2025. Prior to this commit, scuffle was going to be a `Live
Streaming Platform` similar to Twitch or YouTube. The project's current
goal is to create an open-source cloud provider similar to Cloudflare,
Mux, AWS or GCP. We want to provide cost effective services that can be
self-hosted easily with minimal configuration, and easy interoperability
between various other providers and our services.

Our first product we will launch will be video streaming services for
static or live content.

This commit also relicences the project; moving from BSL to a AGPL3 for
applications. MIT & Apache 2 for all library code.

Remaining Refactor Work:

Currently the ./video folder has a lot of libraries and code that we
have not yet ported to use our new library & binary design.